\name{getExtension}
\alias{getExtension}
\title{
  Get File Extension
}
\description{
  Helper function used internally to get file extension.
}
\usage{
  getExtension(filename)
}
\arguments{
  \item{filename}{
    Character
  }
}
\value{
  Character with file extension
}
\keyword{internal}
\author{
  Vincent T van Hees <v.vanhees@accelting.com>
}